% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_collapseTableByLevenSim.R
\name{helper_collapseTableByLevenSim}
\alias{helper_collapseTableByLevenSim}
\title{helper_collapseTableByLevenSim}
\usage{
helper_collapseTableByLevenSim(inputTable, levenSimThresholdVal)
}
\arguments{
\item{inputTable}{dataframe. HOMER output table modified in the parent script--ready for filtering by Levenshtein similarity.}

\item{levenSimThresholdVal}{float. Value for thresholding TFs. For groups of TFs with similar consensus sequences, the TF with the lowest p-value by HOMER will be retained.}
}
\value{
tibble
}
\description{
Reads in a table and value for Levenshtein threshold and returns a table collapsed by threshold (highest p-value for each group)
}
\author{
Tim Scott
}
