% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_getMaxLevenSimCol.R
\name{helper_getMaxLevenSimCol}
\alias{helper_getMaxLevenSimCol}
\title{helper_getMaxLevenSimCol}
\usage{
helper_getMaxLevenSimCol(vectorOfMotifs)
}
\arguments{
\item{vectorOfMotifs}{vector of char. Vector of motifs to filter through.}
}
\value{
data.frame
}
\description{
Reads in a vector of motifs and returns a
}
\author{
Tim Scott
}
