% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_makeBigTableFromListOfTables.R
\name{helper_makeBigTableFromListOfStandardTables}
\alias{helper_makeBigTableFromListOfStandardTables}
\title{makeBigTableFromListofTables}
\usage{
helper_makeBigTableFromListOfStandardTables(inputListOfTables)
}
\arguments{
\item{inputListOfTables}{list of dataframe. List of HOMER TF knownResults Tables.}
}
\value{
list of tibble
}
\description{
Reads in a list of tables and return list of tables with percent Fold Change (enrichment)
}
\author{
Tim Scott
}
